+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/manage/ManageForm.tsx
@@ -1,10 +1,11 @@
 "use client";
-import { FormOwnership } from "@formBuilder/components";
+
 import { FormRecord } from "@lib/types";
 import { useSession } from "next-auth/react";
 import { useRehydrate } from "@lib/hooks/form-builder";
 import { DownloadForm } from "./DownloadForm";
 import { SetClosingDate } from "./SetClosingDate";
+import { FormOwnership } from "./FormOwnership";
 
 interface AssignUsersToTemplateProps {
   formRecord?: FormRecord;
