+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/index.ts
@@ -1,3 +1,2 @@
 export { ShareModal } from "./ShareModal";
 export { ShareModalUnauthenticated } from "./ShareModalUnauthenticated";
-export { FormOwnership } from "./FormOwnership";
